+++ AI_Sports/AISports.Entity/TrainingActivityRecordEntity.cs
@@ -30,9 +30,9 @@ namespace AI_Sports.Entity
         /// 主键自增id
         public long Id { get; set; }
         /// 外键训练课程id
-        public int Fk_training_course_id { get; set; }
+        public long Fk_training_course_id { get; set; }
         /// 外键，训练活动ID
-        public int Fk_activity_id { get; set; }
+        public long Fk_activity_id { get; set; }
         ///  训练活动编码：力量循环或者力量耐力循环
         public string Activity_type { get; set; }
         /// 课程轮次计数：等于插入时训练课程表的当前课程计数，标志这条训练活动记录属于第几次的课程
